+++ finagle-base-http/src/main/scala/com/twitter/finagle/http/codec/context/HttpRetriesContext.scala
@@ -1,23 +1,26 @@
 package com.twitter.finagle.http.codec.context
 
-import com.twitter.finagle.context.{Contexts, Retries}
-import com.twitter.util.{Return, Throw, Try}
+import com.twitter.finagle.context.Contexts
+import com.twitter.finagle.context.Requeues
+import com.twitter.util.Return
+import com.twitter.util.Throw
+import com.twitter.util.Try
 import scala.util.control.NonFatal
 
-private object HttpRetries extends HttpContext {
+private object HttpRequeues extends HttpContext {
 
-  type ContextKeyType = Retries
-  val key: Contexts.broadcast.Key[Retries] = Retries
+  type ContextKeyType = Requeues
+  val key: Contexts.broadcast.Key[Requeues] = Requeues
 
-  def toHeader(retries: Retries): String = {
-    retries.attempt.toString
+  def toHeader(requeues: Requeues): String = {
+    requeues.attempt.toString
   }
 
-  def fromHeader(header: String): Try[Retries] = {
+  def fromHeader(header: String): Try[Requeues] = {
     try {
-      Return(Retries(header.toInt))
+      Return(Requeues(header.toInt))
     } catch {
-      case NonFatal(e) => Throw(new NumberFormatException)
+      case NonFatal(_) => Throw(new NumberFormatException)
     }
   }
 }