+++ finagle-base-http/src/test/scala/com/twitter/finagle/http/filter/ContextFilterTest.scala
@@ -2,38 +2,43 @@ package com.twitter.finagle.http.filter
 
 import com.twitter.conversions.DurationOps._
 import com.twitter.finagle.Service
-import com.twitter.finagle.context.{Contexts, Deadline, Retries}
+import com.twitter.finagle.context.Contexts
+import com.twitter.finagle.context.Deadline
+import com.twitter.finagle.context.Requeues
 import com.twitter.finagle.http.codec.context.HttpContext
-import com.twitter.finagle.http.{Request, Response, Status}
-import com.twitter.util.{Await, Future}
+import com.twitter.finagle.http.Request
+import com.twitter.finagle.http.Response
+import com.twitter.finagle.http.Status
+import com.twitter.util.Await
+import com.twitter.util.Future
 import org.scalatest.funsuite.AnyFunSuite
 
 class ContextFilterTest extends AnyFunSuite {
 
   test("parses Finagle-Ctx headers") {
     val writtenDeadline = Deadline.ofTimeout(5.seconds)
-    val writtenRetries = Retries(5)
+    val writtenRequeues = Requeues(5)
     val service =
       new ClientContextFilter[Request, Response] andThen
         new ServerContextFilter[Request, Response] andThen
         Service.mk[Request, Response] { req =>
           assert(Deadline.current.get == writtenDeadline)
-          assert(Retries.current.get == writtenRetries)
+          assert(Requeues.current.get == writtenRequeues)
           Future.value(Response())
         }
 
     Contexts.broadcast.let(Deadline, writtenDeadline) {
-      Contexts.broadcast.let(Retries, writtenRetries) {
+      Contexts.broadcast.let(Requeues, writtenRequeues) {
         val req = Request()
         HttpContext.write(req)
 
-        // Clear the deadline/retries values in the context
+        // Clear the deadline/requeues values in the context
         Contexts.broadcast.letClearAll {
           // ensure the deadline was cleared
           assert(Deadline.current == None)
 
-          // ensure the retries was cleared
-          assert(Retries.current == None)
+          // ensure the requeues was cleared
+          assert(Requeues.current == None)
 
           val rsp = Await.result(service(req))
           assert(rsp.status == Status.Ok)